fixed bug that would lock player in movement after a dialog
The problem was the "if not len(self.game.text_dialog_queue) != 0" check in the
event checks. If the player was moving while initiation dialog and let go of
the movement key during the dialog the "stop moving"-command would not be
recorded and so the player would have a permanent extra velocity in the way
he was moving when initiation the dialog.

The solution was simply to remove this check. It does not matter if the
movement keys are recorded during dialog, because the player will not be
updated anyway.